@@ -1,44 +1,45 @@
 buildscript {
     repositories {
         maven { url = "https://files.minecraftforge.net/maven" }
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath group: "net.minecraftforge.gradle", name: "ForgeGradle", version: "3.+", changing: true
+        classpath group: "net.minecraftforge.gradle", name: "ForgeGradle", version: '5.1.+', changing: true
     }
+}
+
+plugins {
+    id 'java'
 }
 apply plugin: "net.minecraftforge.gradle"
 
+// these two are required for the java plugin to generate jar files with a version
 version = "${version_major}.${version_minor}.${version_patch}"
 group = "mezz.itemzoom" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
-archivesBaseName = "itemzoom-" + mcversion
 
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8 // Need this here so eclipse task generates correctly.
-compileJava {
-    sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
-}
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
 
 minecraft {
-    mappings channel: "snapshot", version: project.mcp_mappings
+    mappings channel: 'official', version: project.minecraft_version
 
     accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
 
     runs {
         client {
-            properties 'forge.logging.console.level': 'debug'
+            property 'forge.logging.console.level', 'debug'
             workingDirectory project.file('run')
             mods {
                 itemzoom {
-                    source sourceSets.main
+                    sources sourceSets.main
                 }
             }
         }
         server {
-            properties 'forge.logging.console.level': 'debug'
+            property 'forge.logging.console.level', 'debug'
             workingDirectory project.file('run')
             mods {
                 itemzoom {
-                    source sourceSets.main
+                    sources sourceSets.main
                 }
             }
         }
