@@ -2,17 +2,16 @@ package mezz.itemzoom.client.compat;
 
 import java.util.Optional;
 
+import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.runtime.IIngredientListOverlay;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.ItemStack;
 
 public class JeiCompat {
 	public static Optional<IIngredientListOverlay> ingredientListOverlay = Optional.empty();
 
 	public static ItemStack getStackUnderMouse() {
 		return ingredientListOverlay
-				.map(IIngredientListOverlay::getIngredientUnderMouse)
-				.filter(ingredient -> ingredient instanceof ItemStack)
-				.map(ingredient -> (ItemStack) ingredient)
+				.map(i -> i.getIngredientUnderMouse(VanillaTypes.ITEM))
 				.orElse(ItemStack.EMPTY);
 	}
 
