@@ -1,12 +1,9 @@
 package mezz.itemzoom.client.config;
 
 import mezz.itemzoom.client.compat.JeiCompat;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.ForgeConfigSpec.ConfigValue;
 
-@OnlyIn(Dist.CLIENT)
 public class Config {
 	private static final int MIN_ZOOM = 10;
 	private static final int MAX_ZOOM = 100;
@@ -16,8 +13,9 @@ public class Config {
 	private final ConfigValue<Integer> zoomAmount;
 	private final ConfigValue<Boolean> jeiOnly;
 	private final ConfigValue<Boolean> showHelpText;
-	private final ConfigValue<Boolean> showDamageBar;
+	private final ConfigValue<Boolean> showDurabilityBar;
 	private final ConfigValue<Boolean> showStackSize;
+	private final ConfigValue<Boolean> showCooldown;
 	private final ForgeConfigSpec configSpec;
 
 	public Config() {
@@ -44,7 +42,7 @@ public class Config {
 				.translation("config.itemzoom.show.help.text")
 				.define("show.help.text", true);
 
-		showDamageBar = builder
+		showDurabilityBar = builder
 				.comment("Display the item's durability bar when zoomed.")
 				.translation("config.itemzoom.show.damage.bar")
 				.define("show.damage.bar", false);
@@ -53,6 +51,11 @@ public class Config {
 				.comment("Display the item's stack size when zoomed.")
 				.translation("config.itemzoom.show.stack.size")
 				.define("show.stack.size", false);
+
+		showCooldown = builder
+				.comment("Display the item's cooldown animation when zoomed.")
+				.translation("config.itemzoom.show.cooldown")
+				.define("show.cooldown", false);
 
 		configSpec = builder.build();
 	}
@@ -84,12 +87,16 @@ public class Config {
 		return showHelpText.get();
 	}
 
-	public boolean showDamageBar() {
-		return showDamageBar.get();
+	public boolean showDurabilityBar() {
+		return showDurabilityBar.get();
 	}
 
 	public boolean showStackSize() {
 		return showStackSize.get();
+	}
+
+	public boolean showCooldown() {
+		return showCooldown.get();
 	}
 
 	public void setZoomAmount(int zoomAmount) {
